refactor(negative): move constants to constant.py
