+++ e2e/libs/kubelet/kubelet.py
@@ -4,7 +4,8 @@ from workload.pod import create_pod
 from workload.pod import delete_pod
 from workload.pod import new_pod_manifest
 
-from workload.pod import IMAGE_UBUNTU
+from workload.constant import IMAGE_UBUNTU
+
 
 def restart_kubelet(node_name, stop_time_in_sec=10):
     manifest = new_pod_manifest(
