+++ e2e/libs/node/constant.py
@@ -1,0 +1,6 @@
+NODE_STRESS_CPU_LOAD_PERCENTAGE = 100
+NODE_STRESS_MEM_LOAD_PERCENTAGE = 100
+NODE_STRESS_MEM_VM_WORKERS = 1
+NODE_STRESS_TIMEOUT_SECOND = 300
+
+LABEL_STRESS_HELPER = "longhorn-stress-helper"
