+++ e2e/libs/node_exec/constant.py
@@ -1,0 +1,2 @@
+DEFAULT_POD_TIMEOUT = 180
+DEFAULT_POD_INTERVAL = 1
