+++ e2e/libs/recurring_job/constant.py
@@ -1,0 +1,2 @@
+RETRY_COUNTS = 180
+RETRY_INTERVAL = 1
