+++ e2e/libs/replica/constant.py
@@ -1,0 +1,2 @@
+RETRY_COUNTS = 150
+RETRY_INTERVAL = 1
