+++ e2e/libs/volume/constant.py
@@ -1,0 +1,11 @@
+KIBIBYTE = 1024
+MEBIBYTE = (KIBIBYTE * KIBIBYTE)
+GIBIBYTE = (MEBIBYTE * KIBIBYTE)
+
+RETRY_COUNTS = 150
+RETRY_INTERVAL = 1
+
+VOLUME_FRONTEND_BLOCKDEV = "blockdev"
+VOLUME_FRONTEND_ISCSI = "iscsi"
+
+DEV_PATH = "/dev/longhorn/"
