+++ e2e/libs/volume/crd.py
@@ -8,10 +8,7 @@ from utility.utility import logging
 from volume.base import Base
 from volume.rest import Rest
 
-
-Ki = 2**10
-Mi = 2**20
-Gi = 2**30
+from volume.constant import GIBIBYTE
 
 class CRD(Base):
 
@@ -38,7 +35,7 @@ class CRD(Base):
             "spec": {
                 "frontend": "blockdev",
                 "replicaAutoBalance": "ignored",
-                "size": str(int(size) * Gi),
+                "size": str(int(size) * GIBIBYTE),
                 "numberOfReplicas": int(replica_count)
             }
         }
