+++ e2e/libs/workload/constant.py
@@ -1,0 +1,3 @@
+IMAGE_BUSYBOX = 'busybox:1.34.0'
+IMAGE_LITMUX = 'litmuschaos/go-runner:latest'
+IMAGE_UBUNTU = 'ubuntu:16.04'
