+++ e2e/libs/workload/pod.py
@@ -6,10 +6,7 @@ from utility.utility import logging
 from utility.utility import generate_name
 from utility.utility import get_retry_count_and_interval
 
-
-IMAGE_BUSYBOX = 'busybox:1.34.0'
-IMAGE_LITMUX = 'litmuschaos/go-runner:latest'
-IMAGE_UBUNTU = 'ubuntu:16.04'
+from workload.constant import IMAGE_BUSYBOX
 
 def new_pod_manifest(image="", command=[], args=[],
                      claim_name="", node_name="", labels={}):
